requirements.txt
# Paper Research Tool — Dependencies
pyyaml>=6.0
